maybe we can reserve the Object and ComplexObject macros for resolving edges and complex data resolution, and simple conversions between data types can be encapsulated in a From trait.
