+++ crates/sui-graphql-rpc/src/types/object.rs
@@ -3,6 +3,8 @@
 
 use async_graphql::dataloader::{DataLoader, LruCache};
 use async_graphql::{connection::Connection, *};
+use sui_json_rpc_types::SuiRawData;
+use sui_sdk::types::object::Owner as NativeOwner;
 
 use super::big_int::BigInt;
 use super::name_service::NameService;
@@ -14,16 +16,52 @@ use super::{
 use crate::server::sui_sdk_data_provider::SuiClientLoader;
 use crate::{server::context_ext::DataProviderContextExt, types::base64::Base64};
 
-#[derive(Clone, Eq, PartialEq, Debug)]
+#[derive(Clone, Eq, PartialEq, Debug, SimpleObject)]
+#[graphql(complex)]
 pub(crate) struct Object {
     pub address: SuiAddress,
     pub version: u64,
     pub digest: String,
     pub storage_rebate: Option<BigInt>,
-    pub owner: Option<SuiAddress>,
+    pub owner: Option<Owner>,
     pub bcs: Option<Base64>,
     pub previous_transaction: Option<TransactionDigest>,
     pub kind: Option<ObjectKind>,
+}
+
+impl From<&sui_json_rpc_types::SuiObjectData> for Object {
+    fn from(s: &sui_json_rpc_types::SuiObjectData) -> Self {
+        Self {
+            version: s.version.into(),
+            digest: s.digest.to_string(),
+            storage_rebate: s.storage_rebate.map(BigInt::from),
+            address: SuiAddress::from_array(**s.object_id),
+            owner: s
+                .owner
+                .unwrap()
+                .get_owner_address()
+                .map(|x| SuiAddress::from_array(x.to_inner()))
+                .ok()
+                .map(|q| Owner { address: q }),
+            bcs: s.bcs.as_ref().map(|raw| match raw {
+                SuiRawData::Package(raw_package) => {
+                    Base64::from(bcs::to_bytes(raw_package).unwrap())
+                }
+                SuiRawData::MoveObject(raw_object) => Base64::from(&raw_object.bcs_bytes),
+            }),
+            previous_transaction: Some(TransactionDigest::from_array(
+                s.previous_transaction.unwrap().into_inner(),
+            )),
+            kind: Some(match s.owner.unwrap() {
+                NativeOwner::AddressOwner(_) => ObjectKind::Owned,
+                NativeOwner::ObjectOwner(_) => ObjectKind::Child,
+                NativeOwner::Shared {
+                    initial_shared_version: _,
+                } => ObjectKind::Shared,
+                NativeOwner::Immutable => ObjectKind::Immutable,
+            }),
+        }
+    }
 }
 
 #[derive(Enum, Copy, Clone, Eq, PartialEq, Debug)]
@@ -53,24 +91,8 @@ pub(crate) struct ObjectKey {
 
 #[allow(unreachable_code)]
 #[allow(unused_variables)]
-#[Object]
+#[ComplexObject]
 impl Object {
-    async fn version(&self) -> u64 {
-        self.version
-    }
-
-    async fn digest(&self) -> String {
-        self.digest.clone()
-    }
-
-    async fn storage_rebate(&self) -> Option<BigInt> {
-        self.storage_rebate.clone()
-    }
-
-    async fn bcs(&self) -> Option<Base64> {
-        self.bcs.clone()
-    }
-
     async fn previous_transaction_block(
         &self,
         ctx: &Context<'_>,
@@ -81,14 +103,6 @@ impl Object {
         } else {
             Ok(None)
         }
-    }
-
-    async fn kind(&self) -> Option<ObjectKind> {
-        self.kind
-    }
-
-    async fn owner(&self) -> Option<Owner> {
-        self.owner.as_ref().map(|q| Owner { address: *q })
     }
 
     // =========== Owner interface methods =============
